Use Jekyll
The CSS for my home-grown static site generator doesn't look that great
and I don't really want to maintain it going forward.

This uses github pages to build the site which is much simpler.

I still need to find a way to move elo_rating/ out of here to its own
project but not sure how to yet given the custom domain issue:
https://github.com/isaacs/github/issues/547